+++ AWSS3/AWSS3TransferUtility.m
@@ -18,6 +18,7 @@
 #import "AWSS3Service.h"
 #import "AWSS3TransferUtilityDatabaseHelper.h"
 #import "AWSS3TransferUtilityTasks.h"
+#import "AWSS3CreateMultipartUploadRequest+RequestHeaders.h"
 
 #import <AWSCore/AWSFMDB.h>
 #import <AWSCore/AWSSynchronizedMutableDictionary.h>
@@ -228,9 +229,6 @@ static int const AWSS3TransferUtilityMultiPartDefaultConcurrencyLimit = 5;
 @end
 
 @interface AWSS3TransferUtility (HeaderHelper)
--(void) propagateHeaderInformation: (AWSS3CreateMultipartUploadRequest *) uploadRequest
-                        expression: (AWSS3TransferUtilityMultiPartUploadExpression *) expression;
-
 -(void) filterAndAssignHeaders:(NSDictionary<NSString *, NSString *> *) requestHeaders
         getPresignedURLRequest:(AWSS3GetPreSignedURLRequest *) getPresignedURLRequest
                     URLRequest: (NSMutableURLRequest *) URLRequest;
@@ -1327,7 +1325,7 @@ static AWSS3TransferUtility *_defaultS3TransferUtility = nil;
     uploadRequest.bucket = bucket;
     uploadRequest.key = key;
 
-    [self propagateHeaderInformation:uploadRequest expression:transferUtilityMultiPartUploadTask.expression];
+    [AWSS3CreateMultipartUploadRequest propagateHeaderInformation:uploadRequest requestHeaders:transferUtilityMultiPartUploadTask.expression.requestHeaders];
     
     //Initiate the multi part
     return [[self.s3 createMultipartUpload:uploadRequest] continueWithBlock:^id(AWSTask *task) {